// src/Type/validation/contactUsValType.type.ts
export type contactUsValType = {
    valContactUsNameRequir: string;
    ValContactUsNameMin: string;
    ValContactUsNameMax: string;
    valContactUsEmailRequir: string;
    ValContactUsEmailNotValid: string;
    valContactUsPhoneRequir: string;
    ValContactUsPhoneNotValid: string;
    ValContactUsPhoneNotMin: string;
    ValContactUsPhoneNotMax: string;
    valContactUsMessageRequir: string;
    ValContactUsMessageMin: string;
    ValContactUsMessageMax: string;
    contactUsformSuccessMessage: string;
};